Continues building fipa and tac skills, allows for several handlers
